Fix issue with updating package which was deleted manually by user // Resolve #555

+++ platformio/pkgmanager.py
@@ -126,7 +126,8 @@ class PackageManager(object):
             click.secho("Not installed", fg="yellow")
             return False
 
-        rmtree(join(self._package_dir, name))
+        if isdir(join(self._package_dir, name)):
+            rmtree(join(self._package_dir, name))
         self._unregister(name)
         click.echo("[%s]" % click.style("OK", fg="green"))
